+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased](https://github.com/rxRust/rxRust/compare/v0.8.2...HEAD)
 
+### Bug Fixes
+
+- **operator**: `sample` support clone, and not require source observer and sample observable return same subscription.
+
 ## [0.8.2](https://github.com/rxRust/rxRust/releases/tag/v0.8.2)  (2020-03-25)
 
 ### Breaking Changes